+++ lpp/hooks.py
@@ -202,7 +202,9 @@ doc_events = {
 #
 override_whitelisted_methods = {
 	"erpnext.controllers.stock_controller.make_quality_inspections": "lpp.custom.custom_quality_inspection.custom_make_quality_inspections",
-    "erpnext.stock.doctype.purchase_receipt.purchase_receipt.make_purchase_invoice" : "lpp.custom.custom_purchase_receipt.make_purchase_invoice"
+    "erpnext.stock.doctype.purchase_receipt.purchase_receipt.make_purchase_invoice" : "lpp.custom.custom_purchase_receipt.make_purchase_invoice",
+    "thai_tax.custom.custom_api.get_withholding_tax" : "lpp.custom.custom_payment_entry.get_withholding_tax",
+    "thai_tax.custom.custom_api.make_withholding_tax_cert" : "lpp.custom.custom_payment_entry.make_withholding_tax_cert"
 }
 #
 # each overriding function accepts a `data` argument;
