+++ lpp/public/js/payment_entry.js
@@ -12,77 +12,6 @@ frappe.ui.form.on("Payment Entry", {
                 filters: { name: ["in", doctypes] },
             };
         });
-
-        setTimeout(() => {
-            if (frm.custom_buttons && frm.custom_buttons[('Create Withholding Tax Cert')]) {
-                // Remove Original Button
-                frm.remove_custom_button(__('Create Withholding Tax Cert'))
-                frm.add_custom_button(__("Create Withholding Tax Cert"), async function () {
-                    let income_tax_form = "";
-                    if (frm.doc.party_type == "Supplier") {
-                        income_tax_form = (
-                            await frappe.db.get_value(
-                                frm.doc.party_type, frm.doc.party, "custom_default_income_tax_form"
-                            )
-                        ).message.custom_default_income_tax_form;
-                    }
-                    const fields = [
-                        {
-                            fieldtype: "Link",
-                            label: __("WHT Type"),
-                            fieldname: "wht_type",
-                            options: "Withholding Tax Type",
-                            default: frm.doc.custom_wht_type,
-                            reqd: 1,
-                        },
-                        {
-                            fieldtype: "Date",
-                            label: __("Date"),
-                            fieldname: "date",
-                            reqd: 1,
-                        },
-                        {
-                            fieldtype: "Select",
-                            label: __("Income Tax Form"),
-                            fieldname: "income_tax_form",
-                            options: "PND3\nPND53",
-                            default: income_tax_form
-                        },
-                        {
-                            fieldtype: "Link",
-                            label: __("Company Address"),
-                            fieldname: "company_address",
-                            options: "Address",
-                            get_query: () => {
-                                return {
-                                    filters: {
-                                        is_your_company_address: 1,
-                                    },
-                                };
-                            },
-                        },
-                    ];
-                    frappe.prompt(
-                        fields,
-                        function (filters) {
-                            frm.events.make_withholding_tax_cert(frm, filters);
-                        },
-                        __("Withholding Tax Cert"),
-                        __("Create Withholding Tax Cert")
-                    );
-                });
-            }
-            // ตรวจสอบว่าปุ่มถูกสร้างแล้วหรือยัง
-            if (frm.fields_dict['deduct_withholding_tax'] && frm.fields_dict['deduct_withholding_tax'].input) {
-                // ยกเลิกการผูกฟังก์ชันเดิมของปุ่มที่สร้างจาก Customize Form
-                $(frm.fields_dict['deduct_withholding_tax'].input).off('click');
-
-                // ผูกฟังก์ชันใหม่กับปุ่ม
-                $(frm.fields_dict['deduct_withholding_tax'].input).on('click', function() {
-                    frm.trigger('deduct_withholding_tax_lpp');
-                });
-            }
-        }, 10);
     },
     validate(frm) {  
         if (frm.doc.references){              
@@ -95,56 +24,6 @@ frappe.ui.form.on("Payment Entry", {
             }
         });
         }
-    },
-    // Called from button Duduct Withholding Tax
-	deduct_withholding_tax_lpp: function (frm) {
-		const fields = [
-			{
-				fieldtype: "Link",
-				label: __("WHT Type"),
-				fieldname: "wht_type",
-				options: "Withholding Tax Type",
-				reqd: 1,
-			},
-		];
-		frappe.prompt(
-			fields,
-			function (filters) {
-				frm.events.add_withholding_tax_deduction(frm, filters);
-				frm.doc.custom_wht_type = filters.wht_type
-			},
-			__("Deduct Withholding Tax"),
-			__("Add Withholding Tax Deduction")
-		);
-	},
-	add_withholding_tax_deduction: function (frm, filters) {
-		return frappe.call({
-			method: "lpp.custom.payment_entry.get_withholding_tax",
-			args: {
-				filters: filters,
-				doc: frm.doc,
-			},
-			callback: function (r) {
-				var d = frm.add_child("deductions");
-				d.account = r.message["account"];
-				d.cost_center = r.message["cost_center"];
-				d.amount = r.message["amount"];
-				frm.refresh();
-			},
-		});
-	},
-    make_withholding_tax_cert: function (frm, filters) {
-        return frappe.call({
-            method: "lpp.custom.payment_entry.make_withholding_tax_cert",
-            args: {
-                filters: filters,
-                doc: frm.doc,
-            },
-            callback: function (r) {
-                var doclist = frappe.model.sync(r.message);
-                frappe.set_route("Form", doclist[0].doctype, doclist[0].name);
-            },
-        });
     },
 });
